flytter ut migrering i egen klasse for å ta i bruk transactional
